fix: Remove outdated languages that are blocking the real translations
Currently there are some translations that do not exist on transifex,
but because they share the same name as the real language (like `el_GR` vs `el`)
they are used instead.

In this example `el_GR` was loaded instead of `el` but the real translations are in `el`.

I compared the list of languages with transifex.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>