fix: Always request current content when triggering a menu entry
We need to make sure we work with an up-to-date version of the current content,
e.g. if a file was removed or added, so reload the current content on click.

The using app (esp. files) should have cached version so no requests are needed.
(for files this is already the case).

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>